sdk: extract owned post-registration state machine
