+++ crates/walrus-sdk/src/node_client.rs
@@ -84,12 +84,12 @@ use crate::{
         auto_tune::AutoTuneHandle,
         byte_range_read_client::ByteRangeReadClient,
         client_types::{
-            BlobAwaitingUpload,
             BlobData,
-            BlobPendingCertifyAndExtend,
             BlobWithStatus,
             EncodedBlob,
-            RegisteredBlob,
+            OwnedBlobAwaitingUpload,
+            OwnedBlobPendingCertifyAndExtend,
+            OwnedRegisteredBlob,
             UnencodedBlob,
             WalrusStoreBlobFinished,
             WalrusStoreBlobMaybeFinished,
